+++ packages/events/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @goodie-ts/events
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [4ca51c5]
+  - @goodie-ts/core@0.7.0
+
 ## 0.6.0
 
 ### Minor Changes